Refactored sections and components.
