+++ src/App.css
@@ -5,6 +5,8 @@
   --CL-primary-blue-25: #3858d6e8;
   --CL-secondary: #f6f6f6;
   --CL-secondary-darker: #8b8e99;
+
+  --CL-light: #e0e0e0;
   --CL-neutral-white: #fff;
   --CL-neutral-black: #000;
   --FF-primary: 'Poppins', sans-serif;
