+++ src/sections/Products/Products.css
@@ -23,8 +23,7 @@
   padding: 2.3em 0;
   display: grid;
   place-items: center;
-  gap: 1em;
-  row-gap: 3em;
+  gap: 3em;
 }
 
 @media (min-width: 950px) {
